Reorganize ejson module, moving types into files.
