+++ ejson/src/main/scala/quasar/ejson/package.scala
@@ -17,7 +17,6 @@
 package quasar
 
 import slamdata.Predef._
-import quasar.contrib.matryoshka._
 import quasar.fp.ski._
 
 import java.lang.String
@@ -26,10 +25,8 @@ import scala.Unit
 import scala.collection.immutable.{List, ListMap}
 import scala.math.{BigDecimal, BigInt}
 
-import matryoshka._
-import matryoshka.implicits._
 import monocle.{Iso, Prism}
-import scalaz._, Scalaz._
+import scalaz._
 
 package object ejson {
   def arr[A] =
@@ -76,133 +73,4 @@ package object ejson {
   type EJson[A]   = Coproduct[Extension, Common, A]
   val ExtEJson    = implicitly[Extension :<: EJson]
   val CommonEJson = implicitly[Common :<: EJson]
-
-  object EJson {
-    def fromJson[A](f: String => A): Json[A] => EJson[A] =
-      json => Coproduct(json.run.leftMap(Extension.fromObj(f)))
-
-    def fromCommon[T](c: Common[T])(implicit T: Corecursive.Aux[T, EJson]): T =
-      CommonEJson(c).embed
-
-    def fromExt[T](e: Extension[T])(implicit T: Corecursive.Aux[T, EJson]): T =
-      ExtEJson(e).embed
-
-    def arr[T[_[_]]](xs: T[EJson]*)(implicit T: CorecursiveT[T]): T[EJson] =
-      fromCommon(Arr(xs.toList))
-
-    def bool[T[_[_]]](b: Boolean)(implicit T: CorecursiveT[T]): T[EJson] =
-      fromCommon(Bool(b))
-
-    def dec[T[_[_]]](d: BigDecimal)(implicit T: CorecursiveT[T]): T[EJson] =
-      fromCommon(Dec(d))
-
-    def nul[T[_[_]]](implicit T: CorecursiveT[T]): T[EJson] =
-      fromCommon(Null())
-
-    def str[T[_[_]]](s: String)(implicit T: CorecursiveT[T]): T[EJson] =
-      fromCommon(Str(s))
-
-    def int[T[_[_]]](d: BigInt)(implicit T: CorecursiveT[T]): T[EJson] =
-      fromExt(ejson.Int(d))
-
-    def obj[T[_[_]]](xs: (String, T[EJson])*)(implicit T: CorecursiveT[T]): T[EJson] =
-      map((xs.map { case (s, t) => str[T](s) -> t }): _*)
-
-    def map[T[_[_]]](xs: (T[EJson], T[EJson])*)(implicit T: CorecursiveT[T]): T[EJson] =
-      fromExt(ejson.Map(xs.toList))
-
-    def isNull[T](ej: T)(implicit T: Recursive.Aux[T, EJson]): Boolean =
-      CommonEJson.prj(ej.project) exists (ejson.nul.nonEmpty(_))
-
-    /** Replaces `Meta` nodes with their value component. */
-    def elideMetadata[T](
-      implicit T: Recursive.Aux[T, EJson]
-    ): EJson[T] => EJson[T] = totally {
-      case ExtEJson(Meta(v, _)) => v.project
-    }
-
-    /** Replace a string with an array of characters. */
-    def replaceString[T](
-      implicit T: Corecursive.Aux[T, EJson]
-    ): EJson[T] => EJson[T] = totally {
-      case CommonEJson(Str(s)) => CommonEJson(ejson.arr[T](s.toList map (c => fromExt(ejson.char[T](c)))))
-    }
-
-    /** Replace an array of characters with a string. */
-    def restoreString[T](
-      implicit
-      TC: Corecursive.Aux[T, EJson],
-      TR: Recursive.Aux[T, EJson]
-    ): EJson[T] => EJson[T] = totally {
-      case a @ CommonEJson(Arr(t :: ts)) =>
-        (t :: ts).traverse(t => ExtEJson.prj(t.project) >>= (ejson.char[T].getOption(_)))
-          .fold(a)(cs => CommonEJson(ejson.str[T](cs.mkString)))
-    }
-  }
-
-
-  final case class TypeTag(value: String) extends scala.AnyVal
-
-  object TypeTag {
-    val Binary    = TypeTag("_ejson.binary")
-    val Date      = TypeTag("_ejson.date")
-    val Interval  = TypeTag("_ejson.interval")
-    val Time      = TypeTag("_ejson.time")
-    val Timestamp = TypeTag("_ejson.timestamp")
-
-    val stringIso: Iso[TypeTag, String] =
-      Iso[TypeTag, String](_.value)(TypeTag(_))
-
-    implicit val encodeEJson: EncodeEJson[TypeTag] =
-      EncodeEJson[String].contramap(_.value)
-
-    implicit val order: Order[TypeTag] =
-      Order.orderBy(_.value)
-
-    implicit val show: Show[TypeTag] =
-      Show.showFromToString
-  }
-
-  object Type {
-    import EJson._
-
-    val TypeKey = "_ejson.type"
-
-    def apply[T](tag: TypeTag)(implicit T: Corecursive.Aux[T, EJson]): T =
-      fromExt(Map(List(fromCommon[T](Str(TypeKey)) -> fromCommon[T](Str(tag.value)))))
-
-    /** Extracts the type tag from a metadata map, if present. */
-    def unapply[T](ejs: EJson[T])(implicit T: Recursive.Aux[T, EJson]): Option[TypeTag] =
-      ejs match {
-        case ExtEJson(Map(xs)) =>
-          xs collectFirst {
-            case (Embed(CommonEJson(Str(TypeKey))), Embed(CommonEJson(Str(t)))) => TypeTag(t)
-          }
-        case _ => none
-      }
-  }
-
-  object SizedType {
-    import EJson._
-
-    val SizeKey = "_ejson.size"
-
-    def apply[T](tag: TypeTag, size: BigInt)(implicit T: Corecursive.Aux[T, EJson]): T =
-      fromExt(Map(List(
-        fromCommon[T](Str(Type.TypeKey)) -> fromCommon[T](Str(tag.value)),
-        fromCommon[T](Str(SizeKey))      -> fromExt[T](Int(size))
-      )))
-
-    /** Extracts the type tag and size from a metadata map, if both are present. */
-    def unapply[T](ejs: EJson[T])(implicit T: Recursive.Aux[T, EJson]): Option[(TypeTag, BigInt)] =
-      ejs match {
-        case ExtEJson(Map(xs)) =>
-          val size = xs collectFirst {
-            case (Embed(CommonEJson(Str(SizeKey))), Embed(ExtEJson(Int(s)))) => s
-          }
-          Type.unapply(ejs) tuple size
-
-        case _ => none
-      }
-  }
 }
